+++ deploy/etc/.gitignore
@@ -1,0 +1,3 @@
+database-dump.sql
+*.ini
+schema-local.sql
