+++ deploy/etc/NOTES.md
@@ -1,0 +1,53 @@
+ffmpeg downloaded from https://johnvansickle.com/ffmpeg/
+
+add database comments for each table.
+
+- primary_course_id remembers current course.  -> default_course_id
+
+
+
+- upload page should set cookie
+- list page should set cooike
+
+course edit page:
+- Shows students enrolled
+- Shows the key
+- Shows enrollments left
+- Allows changing the key
+- Shown to all users who are admins for that course.
+
+
+- Need to create superadmin
+- Can create new courses
+- Add admin capabilities to an existing user.
+
+table to implement
+- Log table
+
+
+movie data:
++ Have URL for a movie instead of blob..
++ Movie_data
+
+movies table - add more fields:
+- Version Number
+- Derrived from
+- Processed
+- Processor
+
+
+analysis:
+movie_analysis
+- Movie_id -
+- analysis_engine id
+- JSON
+
+movie_frame_analysis
+- Movie_id
+- movie_frame_id
+- analysis_engine id
+- JSON
+
+engine table:
+id -
+engine_name - name
