+++ deploy/etc/database_name
@@ -1,0 +1,2 @@
+#!/bin/bash
+grep mysql_database $1 | awk '{print $3;}' | head -1
