+++ deploy/etc/sam-idea-1.yaml
@@ -1,0 +1,38 @@
+AWSTemplateFormatVersion: '2010-09-09'
+Transform: AWS::Serverless-2016-10-31
+Description: >
+  Creates an S3 bucket and a DynamoDB table named based on the username.
+
+Parameters:
+  UserName:
+    Type: String
+    Description: The username to include in resource names.
+
+Resources:
+  UserS3Bucket:
+    Type: AWS::S3::Bucket
+    Properties:
+      BucketName: !Sub "s3-bucket-${UserName}"
+    DeletionPolicy: Delete  # Ensure the bucket is deleted when the stack is removed
+
+  UserDynamoDBTable:
+    Type: AWS::DynamoDB::Table
+    Properties:
+      TableName: !Sub "dynamodb-table-${UserName}"
+      AttributeDefinitions:
+        - AttributeName: ID
+          AttributeType: S
+      KeySchema:
+        - AttributeName: ID
+          KeyType: HASH
+      BillingMode: PAY_PER_REQUEST
+    DeletionPolicy: Delete  # Ensure the table is deleted when the stack is removed
+
+Outputs:
+  S3BucketName:
+    Description: The name of the S3 bucket
+    Value: !Ref UserS3Bucket
+
+  DynamoDBTableName:
+    Description: The name of the DynamoDB table
+    Value: !Ref UserDynamoDBTable
