--- etc/.gitignore
@@ -1,3 +1,0 @@
-database-dump.sql
-*.ini
-schema-local.sql
