--- etc/config-vm.bash
@@ -1,3 +1,0 @@
-snap install emacs --classic
-apt install mysql-server -y
-apt install net-tools
