--- etc/credentials_template.ini
@@ -1,26 +1,0 @@
-# [client] is read by the mysql command line with:
-# mysql --defaults-extra-file=credentials.ini
-# The [client] user needs to be able to create database schemas
-[client]
-user = <user that can make schemas, such as root>
-password = <password for that user>
-host = <host where the database is running>
-database = <database name from below, or sys>
-
-# [s3] your s3 bucket
-[s3]
-s3_bucket = <your s3 bucket>
-
-# [dbreader] requires read-only access to a specific database
-[dbreader]
-mysql_host = <hostname>
-mysql_user = <username>
-mysql_password = <user password>
-mysql_database = <database name>
-
-# [dbwriter] requires read-write access to the specific database
-[dbwriter]
-mysql_host = <hostname>
-mysql_user = <username>
-mysql_password = <user password>
-mysql_database = <database name>
