--- etc/database_name
@@ -1,2 +1,0 @@
-#!/bin/bash
-grep mysql_database $1 | awk '{print $3;}' | head -1
