--- etc/ffmpeg-6.1-amd64-static
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:a0082b064cc83f5606554fa2cc5b07194ade90f6669b1fcfd6499b29861ca403
-size 78937568
