--- etc/known_hosts
@@ -1,3 +1,0 @@
-app.planttracer.com ecdsa-sha2-nistp256 AAAAE2VjZHNhLXNoYTItbmlzdHAyNTYAAAAIbmlzdHAyNTYAAABBBCdyXQQ7Lxji7xiFfvlTdYVqx7PbKC8WDf/sWfRvT1CJdI56UWu6xY7FYEjOp3DSGv3odx7cVAlk08cu7EyIhvg=
-app.planttracer.com ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIHnu/7qzXI8k+fLfe/NcdnmS4cqGK/3i9644W1e1iV8z
-app.planttracer.com ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQCqYSrsvRaDr5BliFV4Qq/B2sv0YZBOMTzN8RSUO4vViANik945MWWP9FI2Jb9cZui9QWG+2JGReVuJg+6lzUQUd5pnJm9ZkrdUHvcOoZcnhzjnJTn6+YRPjKpR+x2ACs7Wj7oFEN0PXnkS3sUGKv8ndBPvGPXcTbtLaFAf2+QHWwXFea+1wb9K98oOYNnGaSuTK5rW1I/PNk3kyg65EO/3JVYjZ1yuDINCrWaf5qJEAqMUZArkajBIqy1g1AwhSfDbNm9tnBe806TlMtjUisqO7/FpqBqCvQHoguKvSYe5KZFG0kRMCNYJW1r6wIqK6b8QyAEi313a0cBe7aPUi5ir
