--- layers/opencv-layer/requirements.txt
@@ -1,1 +1,0 @@
-opencv-python-headless
