--- layers/python-layer/requirements.txt
@@ -1,5 +1,0 @@
-boto3
-flask
-pymysql
-requests
-validate_email_address
